+++ R/read_asm.R
@@ -18,8 +18,7 @@ read_asm <- function(path, data_format = c("wide", "long"),
                      collapse = TRUE){
   format_out <- match.arg(format_out, c("matrix", "data.frame", "data.table"))
   data_format <- check_data_format(data_format, format_out)
-  metadata_format <- match.arg(metadata_format, c("chromconverter", "raw"))
-  metadata_format <- switch(metadata_format, "chromconverter" = "asm", "raw")
+  metadata_format <- check_metadata_format(metadata_format, "asm")
 
   xx <- jsonlite::fromJSON(path, flatten = FALSE)